feat(MonAuto) :adding database

+++ src/main/java/fr/dojo/monauto/data/entity/Client.java
@@ -2,7 +2,7 @@ package fr.dojo.monauto.data.entity;
 
 public class Client {
 
-    private int idClient;
+    private long idClient;
     private String nomClient;
     private String prenomClient;
     private String adresseClient;
